Update the SUPPORT file with troubleshooting steps
This commit updates the `SUPPORT.md` file with troubleshooting steps
for the most commonly seen problems so far.

This change should help in a few ways:
1. Quick self-service for people to help themselves
2. Documentation that the support team at GitHub can reference

The `README.md` file has been updated to link people to the
`SUPPORT.md` file instead of linking directly to the GitHub Support
contact form.