Update `SUPPORT.md` to simplify reinstall steps
This commit updates the `SUPPORT.md` file to simplfy the steps required to reinstall the integration after uninstalling it.

The more complex steps of enabling developer mode and installing from a URL is no longer required after this pull request was merged:

https://github.com/integrations/jira/pull/155

Now all that is required is a reinstall from the Atlassian marketplace.