Updates to address sync getting stuck due to invalid issue keys (#163)
* just log an error on 400s instead of throwing

This is stopping the sync from continuing. These are because of an invalid payload, so we don't need to stop processing, we just need to skip this page of results since it contains bad data.

* this is incorrectly marking issue keys that have an underscore in them

Example:
JSP_JIR-123 matches, but JSP_JIR is not a valid identifier. This is not valid smart-commit syntax, so it's causing bad requests

* update tests to match

* Stringify the payload so we can see it in the logs

* log jiraPayload as an object