Uppercase issue keys
JIRA’s developer info endpoint is strict. If you have a JIRA project that uses `JRA` as th key but submit `jra-123`, it will respond with a 400.

Users want to be able to specify projects in upper or lower case so let’s upper case issue keys before submitting them to JIRA.