Remove support for lower case issue keys
After discussing this with the team and reviewing the Atlassian documentation for Smart Commits and testing out the DVCS Connector, I’ve decided not to add support for lower case issue keys.

There is still some question about whether it would be desirable to have these so I’ve filed #220 to investigate it further.