Fix smart commit parsing bugs
This is a bit of a grab bag commit, including some bug fixes and some clean up. Sorry about that.

Here’s what it includes:

### Fix syntax error due to whitespace

Prior to this change, you could cause Moo to throw a syntax error with some carefully placed whitespace. This was possible because not every state—`main`, `transition`, `workLog`—had a rule for whitespace.

### Fix syntax error due to hash.

Prior to this change, a syntax error would trigger on this string: `Rename Node#move to Node#move_within`. The way we supported multiple commands per line was to blame. To get around this, transition comments are now parsed word-by-word. For example, `JRA-123 #close All finished` is now tokenized like this:

```
[
  {type: 'issueKey', value: 'JRA-123'},
  {type: 'transition', value: 'close'},
  {type: 'whitespace', value: ' '},
  {type: 'comment', value: 'All'},
  {type: 'whitespace', value: ' '},
  {type: 'comment', value: 'finished'},
]
```

Splitting on whitespace makes it easier to detect other more important rules, like transitions or time. This makes the tokenizer more flexible and moves the constructing the comment to `smart-commit.js`.

### Don’t allow commands at the start of a line

The previous Smart Commits implementation allowed this. It even explictly tested this behaviour. According to the Atlassian Smart Commits docs [1], this use case isn’t allowed.

For this reason, and also because it complicated the tokenizer implementation, I’ve removed support for it and modified the tests.

_Note: the reason it doesn’t make sense to start a line with a comment is because Smart Commits is intentionally bound to a single line, which must always include an issue before any command is issued. For example, `JRA-123 #resolve`. Splitting across multiple lines, like `JRA-123\n#resolve`, may result in accidental action being taken on an issue that is mentioned anywhere in a commit body._

[1] https://confluence.atlassian.com/bitbucket/processing-jira-software-issues-with-smart-commit-messages-298979931.html