Remove Smart Commits command parsing
This code isn’t necessary and can be deleted. When we submit commit messages to JIRA, they handle the Smart Commits command parsing. The thing we need to do is extract issue keys from other fields, like branch or pull request title. Now that we have a fully-functional Smart Commits tokenizer, we need only iterate over the tokens and grab the issue keys.

If we decide to apply commands in other fields, like the pull request body, we would need bring this back. However, I think the cost of keeping the code just in case we need it to be too high. Case in point: this last week I spent a non-trivial amount of time ensuring that our code could still parse commands, only to find out later it wasn’t used.