Add ES2018 support to StandardJS
This branch uses regular expression lookbehinds, which aren’t supported in ES2017. We use Node 8.x, which supports them. The proposal is also in stage 4 and included in ES2018.

According to the docs, because the proposal for is stage 4, StandardJS should support it. I’ll file an issue to ask about it but in the meantime, I’ve added `babel-eslint` so we can keep moving.