Improve installation verification
The installation flow for an Atlassian Connect app like this Jira integration has two steps:

1. Install. During this step, Atlassian makes an unauthenticated request to us, providing a shared secret. However, the secret can’t yet be used.
2. Enable. During this step, Atlassian makes an authenticated request using the shared secret, which we verify, and then we make an out-of-band request to their API. If we are successful, we mark the installation as enabled.

We have a handful of installations that appear to be enabled but `enabled == false`. It’s not immediately clear why this is happening but my current hypothesis is that, due to a race condition, our request to Atlassian is never happening because we send the status _before_ adding our callback:

https://github.com/integrations/jira/blob/c903c49eed1a7bfce5e5e2ffece71158b1736f22/lib/jira/enable.js#L7-L8

This pulls the verification into a separate file so we can add a staff endpoint to trigger a verification on behalf of a customer.

[1] Learn about it at https://developer.atlassian.com/cloud/jira/platform/authentication-for-apps/