Add `/api/jira/:id/verify` endpoint
This allows us to trigger Jira installation verification, which finds finds a pending installation (e.g., `enabled=false`) and makes a request to Jira using the shared secret. If that works, it sets `enabled=true`.

This was added so we can retrigger the verification step for failed installations (there are about 87).