Split Sentry from error page middelware
Prior to this commit, the code that added context to Sentry also served up error pages. While both related to errors, these are distinct behaviours that are likely to change separately.

This splits them into two methods, names those methods, and adds those back as middleware.

This is done to ease changes this code in the future.