Don’t catch errors in development
When an error occurs within a request, we display a generic error page to the user. This is useful for production but hides problems in development: instead of seeing a stacktrace or error, you see an Octocat. Cute but not helpful.

This disables the generic error handler middleware when `NODE_ENV=development`. This also adds `NODE_ENV=development` to `.env.example` so folks get this behaviour for free.