Add installation ID to log message
We received an uninstall request from Jira and it appears to have disabled an installation that was still working.

Unfortunately, we can’t be certain of this because the logs don’t include the installation ID. This updates `uninstall.js` to ensure that, should we see this again, we will have confirmation of the problem. As a precautionary measure, I’ve also added installation ID to other log messages.