Remove call to GitHub API from error handler
In order to display a link to the GitHub Apps configuration page (e.g., `https://github.com/apps/jira-jnraine`), we made a call to the GitHub Apps API while handling an error.

Given the frequency of the erorrs and the unreliability of HTTP, this may have been what cause errors to not be delivered upon deploy to production—where error volume is high.

But why wasn’t it a problem before? I have a guess about that. This pull request reordered the error handler to make sure it ran _before_ the error was delivered. I think because it came after, it never interrupted error delivery. Here are the steps:

1. Error raised.
2. Error put into queue to be delivered to Sentry later
3. Error handler to add context starts
4.Request to GitHub Apps API occurs in Promise
5.(Before step 4 finishes) Error pulled from queue and delivered to Sentry