Stop sending error to Sentry multiple times (#239)
When an exception was raised, it was sent to Sentry three or more times. This stops that. It also improves the metadata that comes with each error and stops metadata from leaking between exceptions.

Sentry used to be configured within `lib/error-handler.js` where we called `Sentry.init` and then exported `Sentry`. Anywhere we needed to use Sentry to add context, send errors, or add middleware, we would require `lib/error-handler.js`.

However, `Sentry.init` enables error delivery _globally_ and each time it is called, it appeared to cause errors to be delivered multiple times. We also called `Sentry.captureException` from a function that was called when an exception was raised, which sent the error to Sentry one extra time.

While working on this, I found code that was calling `Sentry.configureScope` instead of `withScope`. The former is intended for metadata that applies to all exceptions but was being called with request-specific metadata. This metadata then leaked across into other exceptions.

This can be tested by [setting SENTRY_DSN](https://forum.sentry.io/t/where-can-i-find-my-dsn/4877/4) locally and then hitting the test endpoints:

- `POST /boom`
- `GET /boom`
- `POST /jira/boom`
- `GET /jira/boom`