Add STALLED sync status
When an active sync has not been updated in the last fifteen minutes, assume something went wrong and mark it as staled. We make a best effort to mark syncs that have failed due to an error as “FAILED” but there are cases where the error handler doesn’t run.

While we work to resolve these issues, the customer shouldn’t need to wait 48 hours before knowing something isn’t working as it should. This helps avoid that.