Throw more specific exceptions (#241)
This is a small tweak to improve the exceptions sent to Sentry.

The first case gives a better idea to the investigating developer. The second case ensures the original exception is sent on to Sentry, including any specifics returned by the GitHub API.

These are minor changes but every bit helps.