Add “Last updated” column
This column appears on the Jira configuration page and aims to help a customer see when a sync was last updated. It is shown as a relative time—e.g., “five minutes ago”.

This also modifies the “Added” column to be relative. In both cases, the absolute time can be seen by hovering over the relative time.