Add STALLED sync status and "Last sync update" column (#240)
When an active sync has not been updated in the last fifteen minutes, assume something went wrong and mark it as staled. We make a best effort to mark syncs that have failed due to an error as “FAILED” but there are cases where the error handler doesn’t run.

While we work to resolve these issues, the customer shouldn’t need to wait 48 hours before knowing something isn’t working as it should. This helps avoid that.

This also adds a "Last sync update" column on the Jira configuration page and aims to help a customer see when a sync was last updated. It is shown as a relative time—e.g., “five minutes ago”.