Improve error handling for webhooks
Prior to this change, the error reporting for webhooks delivered by GitHub and handled by Probot had some shortcomings.

- **The user was identified by IP**, which pointed back to whatever server delivered the webhook. This isn’t meaningful information and made it difficult to tell who was impacted by the error.
- **Errors weren’t searchable.** Related to the first point, the errors had no meaningful tags to search against. I couldn’t filter errors by the type of webhook, the related jira host, or the GitHub installation. Instead, errors were mixed together so long as the exception message looked the same.
- **Error context leaked to other requests.** This is most noticeable by looking at errors in Sentry that shouldn’t have a “GitHub Payload”. If a process sent a webhook error to Sentry and then processed a background job error, the webhook “GitHub Payload” would appear on the background job error in Sentry. This seems to be a bug in the bunyan stream, though I’m not 100% sure.

This leaking goes both ways too. To illustrate this, start the app locally with Sentry enabled, triggering one of the “boom” endpoints, then triggering a webhook error (I did this by adding a throw into my local codebase). When the errors are viewed side-by-side in Sentry, webhook error appears to be sent by curl, which it is not.

To address this, I’ve removed the bunyan stream and added custom error Sentry error handling. This is done in a fashion similar to our background jobs.

With this new error handling in place, exceptions thrown while processing webhook events are tagged with the following data:

- User: jiraHost and gitHubInstallationId
- Tag: jiraHost
- Tag: gitHubInstallationId
- Tag: transaction (includes webhook event and action)
- Extra: A brief summary of the GitHub payload

Should it be necessary, additional context can be added by each event handler by calling `context.sentry.addTag` or `context.sentry.addExtra`.