Make debugging in VS Code easy (#248)
Prior to this commit, I used these steps to debug in VS Code:

1. Add breakpoint
2. Stop server
3. Remove the “web” line from `Procfile.dev`
4. Run `script/server`
5. In VS Code Debugger sidebar, click “Start debugging” (the play icon)

Once I finished debugging, I needed to revert `Procfile.dev` and restart the server. This worked but was a pain. I found myself wishing for something as easy as a browser.

This commit reduces those steps to one: click “Start Debugging” in VS Code.

This is enabled by two changes: the web server is always started with `--inspect` and `.vs_code/launch.json` tells VS Code to look for a running debugger and attach instead of trying to start a server itself. If you want to debug the worker instead of the web process, select “Attach to worker” from the debugger config dropdown and then click the play icon.

I’ll follow-up with some developer documentation to make this accessible to anyone new on the project.