Add stafftools for Jira installation
This allows a staff member to check the authorization status of a specific Jira installation.

```bash
$ bin/stafftools jira 682b4a8f798a480a99e51b04936332a2
{
  "clientKey": "682b4a8f798a480a99e51b04936332a2",
  "host": "https://example.atlassian.net",
  "enabled": true,
  "authorized": true, # <-- Important bit
  "gitHubInstallations": [
    {
      "gitHubInstallationId": 1234,
      "createdAt": "2019-08-26T19:49:41.023Z",
      "updatedAt": "2019-08-26T20:02:33.515Z",
      "syncStatus": “ACTIVE”
    }
  ]
}
```

This is being added to help diagnose 403 exceptions.

As part of this, I’ve introduced a new class, called `JiraClient`. This class does little right now but bit-by-bit, the functionality from `lib/jira/client/index.js` will be migrated into this object. I’ve instead of adding to the existing one to gradually move to a more object oriented style. Along the way, `JiraClient` should be fully tested.