Make the source of the Jira error more obvious
It’s usually not possible to trace an error back to it’s cause using Sentry. If we don’t know how something happened, it’s difficult to fix it.

Right now, the majority of our errors occur while making requests to Jira. When reported, the errors are grouped into three buckets based on the stacktrace.

Unfortunately, all three stacktraces don’t include any mention of our code. Here’s an example of one:

```
Error: Request failed with status code 400
  File "/app/node_modules/axios/lib/core/createError.js", line 16, col 15, in createError
    var error = new Error(message);
  File "/app/node_modules/axios/lib/core/settle.js", line 18, col 12, in settle
    reject(createError(
  File "/app/node_modules/axios/lib/adapters/http.js", line 201, col 11, in IncomingMessage.handleStreamEnd
    settle(resolve, reject, response);
  File "events.js", line 214, col 15, in IncomingMessage.emit
  File "domain.js", line 476, col 20, in IncomingMessage.EventEmitter.emit
  File "_stream_readable.js", line 1178, col 12, in endReadableNT
  File "internal/process/task_queues.js", line 77, col 11, in processTicksAndRejectio
```

Further, it doesn’t matter what path was requested or what the response code was, as long as the stacktrace is the same, it’s grouped together.

This makes it difficult to know what endpoints are most problematic and where in our code the requests originate from.

This commit worked to address this in a few ways:

1. Uses `await` when making the request to `/rest/devinfo/0.10/bulk`. Because we’re running Node 12, this causes our code to appear in the stacktrace. (See details here: https://thecodebarbarian.com/async-stack-traces-in-node-js-12) In order to use `await` with Axios, it has been updated to 0.19.

2. A custom error class, `JiraClientError`, has been created to improve the error message by using a more specific name and including the HTTP method, path and status in the message.

With this change, the stacktrace becomes more meaningful:

```
JiraClientError: POST /rest/devinfo/0.10/bulk1 responded with 404
  File "/Users/jordan/projects/jira/lib/jira/client/axios.js", line 58, col 27, in null.<anonymous>
    return Promise.reject(new JiraClientError(error))
  File "internal/process/task_queues.js", line 85, col 5, in processTicksAndRejections
  File "/Users/jordan/projects/jira/lib/jira/client/index.js", line 168, col 13, in async Object.update
    await instance.post('/rest/devinfo/0.10/bulk1', {
  File "/Users/jordan/projects/jira/lib/github/branch.js", line 13, col 3, in async module.exports.createBranch
    await jiraClient.devinfo.repository.update(jiraPayload)
  File "async /Users/jordan/projects/jira/lib/github/middleware.js", line 53, col 7, in null.<anonymous>
  File "async /Users/jordan/projects/jira/lib/github/middleware.js", line 14, col 7, in null.<anonymous>
```