Escape strings before interpolating into URL paths
This is a precaution to avoid malformed or unexpected URLs.

The Jira client allows issue IDs, repo IDs, or refs to be passed in. Sometimes these values are from the customer. Given that we validate them against some criteria (e.g., must be a valid Jira issue key), it’s unlikely to cause problems but let’s escape the values to be sure.