Continue processing webhooks when error occurs
A customer can connect multiple Jira instances to a GitHub organization. When they do so, GitHub delivers the event as a webhook that we process and fan out to each Jira instance.

If something goes wrong while processing the event for one of the Jira instances, like an HTTP error, we stop processing and throw an error. This can result in events being lost.

To avoid this, we process each event within a `try … catch`. If something goes wrong, we send it to Sentry and continue processing.