refactor mutating functions to make naming more clear?
This commit refactors the issue key checks and mutations by focusing on just being able to call `truncateIssueKeys` or `dedupIssueKeys` from update without caring about any of the internal checks. I also all of the functions for checking issue keys to the bottom of the file since they aren't relevant to most people when they come to the Jira client.