Tweak uninstall to prepare for new stafftools endpoint
In order for uninstalls to be added to stafftools, a few changes need to be made to this function:

1. **Don’t use `Installation.uninstall`.** Because Jira provides an unhashed client key but we store a hashed key, `Installation.uninstall` has built-in hashing. We pass in an unhashed value, it is hashed and then a query is sent to the database with that hashed value. For the stafftools endpoint, we only have the hashed value so we end up hashing it a second time, resulting in a query with no results. Thankfully, we already fetched the installation from the database, so let’s just call `uninstall` on the instance directly. [1]

2. Log after we’re done the work. This is just for tidyness. Why report we’re done before we finish?

I noticed a couple odd things while working in this method. First, the `DELETE`s aren’t wrapped in a transaction. This may lead to half-uninstalled instances. Second, the installation row delete happens after we already send a response to the client. This doesn’t cause a problem, per se, but given the `await`, I’m not sure why this happens.

1: Hashing the value before storing it in the database seems unneccessary. The value is a UUID and on it’s own, can’t be used for anything. It seems like a piece of complexity that we’d be better off without.