Truncate issue keys and show warning to user (#258)
Adds sync warning column to installations and truncates to 100 issue keys. If there are over 100 issue keys referenced, we show a warning to the user. 

This commit refactors the issue key checks and mutations by focusing on just being able to call `truncateIssueKeys` or `dedupIssueKeys` from update without caring about any of the internal checks.

Co-Authored-By: Jordan Raine <jnraine@github.com>