Wait for authorization check
There is a safety in place to avoid accidentally uninstalling an authorized Jira installation.
```
bin/stafftools jira-uninstall 347a4493812968d2c7e674f4d220837b26a9b9b5c0eae709ea5a378ce179938f
{
  "status": "HTTP/1.1 400 Bad Request\r\n",
  "body": "{\"message\":\"Refusing to uninstall authorized Jira installation\"}"
}
```

However, due to a bug, this warning always occurs.

When making the request to test the Jira API shared secret, we don’t wait. This causes `jiraClient.isAuthorized()` to return a promise, which evaluates as a truthy value.

To avoid this, let’s `await` the promise. This ensures the authorization request is run before we proceed with the if statement.