Add Datadog histogram for GitHub requests
When a request is made to the GitHub API using Octokit, a timing histogram is sent to localhost:8125 over UDP. The value of the metric is the elapsed time of the request. Because we’re using a histogram, Datadog will turn this into six different metrics: `count`, `avg`, `median`, `95percentile`, `max`, `min`.

For more information on Datadog histograms, see https://docs.datadoghq.com/developers/metrics/histograms/.

To test this out locally:

1. Start the server
2. In a new tab run `script/datadog-spy`
3. Trigger a sync

You should see metrics printed in the second tab.

Note: Because we don’t have a Datadog agent running on production, _these don’t do anything yet._ That’s okay! This acts as the first step to adding meaning metrics.