Increase value max to 50
I saw a transient failure because the test happened to run slow enough that the value of this metrics was equal to 20.

```
      Error: expect(function)toHaveSentMetrics(metrics)
      - Expected
      + Received
        Array [
          Object {
            "name": "jira-integration.jira_request",
            "tags": Object {
      +       "env": "test",
              "method": "GET",
              "path": "/foo/bar",
              "status": "500",
            },
            "type": "h",
      -     "value": [Function value],
      +     "value": 20,
          },
        ]
```

Tests are liable to run more slowly at times so I’ve increased the max to 50 to ensure this transient failure doesn’t happen again in the future.