Only mark syncs as failed during sync
Customers were reporting syncs switching from COMPLETE to FAILED without triggering a sync themselves. The underlying problem was with how and when we marked syncs as failed.

Syncs occur in the background on the `installation` queue.  When a sync background job throws an error, a global failure handler is called that sets `subscription.syncStatus` to `FAILED`. However, this global failure handler runs when other jobs fail too.

This led to a sync being marked as `FAILED` when any for that customer failed.

This commit remedies that by moving the `subscription.syncStatus` update out of the global failure handler and into the sync job.