Adds hours_since_x for created_at and updated_at
This makes making a histogram in looker much easier with an hours value as part of the view. I tried to push this into the looker model itself, but was running into issues not being able to call postgresql functions.

This also casts the sync_status enum to a text, which also makes our looker model easier to manage.