Add a current command to stafftools
Frequently I find it's not enough to know if a sync is in progress while
debugging, I also often want to know where we are in the process
and how long we have left to go. Also being able to see that things are
changing is helpful to know that a sync is progressing. The new "current"
command shows the cursor for the fetch (if it changes, fetching is
progressing) and the total count of repositories left.