Provide an url when possible on deployment environment
We want to be able to easily access to an environment when possible from
the deployment page.

It needs to be provided as environment_url when making deployment status
call. https://docs.github.com/en/rest/deployments/statuses#create-a-deployment-status

For people who use the classic Github Action to mark deployment as
suggested by Jira. It's here:

https://github.com/chrnorm/deployment-action/blob/880668c419e283adc6b1cc092d52eb5e558fec4d/action.yml#L31-L33

Related:
  - https://github.com/atlassian/github-for-jira/issues/1536
  - https://github.com/atlassian/github-for-jira/pull/1732